refactor: deconstruct single script file to modules. use webpack to avoid eslint errors with imports
