+++ index.html
@@ -8,6 +8,6 @@
   <title>Virtual Keyboard</title>
 </head>
 <body class="body">
-  <script src="./js/script.js"></script>
+  <script src="./scripts.js"></script>
 </body>
 </html>
