+++ package.json
@@ -7,7 +7,8 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "style": "sass --watch sass/style.scss style.css",
     "lint": "eslint ./js",
-    "lint:fix": "eslint ./js --fix"
+    "lint:fix": "eslint ./js --fix",
+    "build": "webpack"
   },
   "author": "Denis Khatsuk",
   "license": "ISC",
@@ -16,6 +17,8 @@
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-plugin-import": "^2.20.1",
-    "sass": "^1.26.3"
+    "sass": "^1.26.3",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11"
   }
 }
